+++ src/app/services/testSuite.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import 'rxjs/add/operator/map';
 import { SimpleRequester } from './simple-requester';
 import { TestSuite, SuiteDashboard } from '../shared/models/testSuite';
+import { Test } from '../shared/models/test';
 
 
 @Injectable()
@@ -49,5 +50,13 @@ export class TestSuiteService extends SimpleRequester {
 
   createSuiteDashboard(suiteDashboard: SuiteDashboard): Promise<SuiteDashboard> {
     return this.doPost(`/suite/dashboard`, suiteDashboard).map(res => res.json()).toPromise();
+  }
+
+  syncSuite(tests: Test[], suiteId: number, removeNotExecutedResults: boolean) {
+    return this.doPost(`/suite/sync`, tests, { suiteId, removeNotExecutedResults }, true).map(res => res).toPromise();
+  }
+
+  findTestToSync(notExecutedFor: number, suiteId: number): Promise<Test[]> {
+    return this.doGet('/suite/sync', { notExecutedFor, suiteId }, true).map(res => res.json()).toPromise();
   }
 }